pyproject.toml
[project]
name = "eval-starter-kit"
version = "0.1.0"
description = "Eval Starter Kit"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.35.0",
    "ollama>=0.5.1",
    "path>=17.1.0",
    "pydash>=8.0.5",
    "fastapi>=0.116.1",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "aioboto3>=15.2.0",
    "mcp>=1.18.0",
    "numpy>=2.3.4",
]

